fix (#1) :: 인터페이스 모듈화
