+++ muldum/src/components/Item/List/data.ts
@@ -1,0 +1,10 @@
+const data = [
+    'AMD 라이젠7 5세대 7800X3D 라파엘 멀티팩 정품 국내정식 유통제품',
+    'AMD 라이젠7 5세대 7800X3D 라파엘 멀티팩 정품 국내정식 유통제품',
+    'AMD 라이젠7 5세대 7800X3D 라파엘 멀티팩 정품 국내정식 유통제품',
+    'AMD 라이젠7 5세대 7800X3D 라파엘 멀티팩 정품 국내정식 유통제품',
+    'AMD 라이젠7 5세대 7800X3D 라파엘 멀티팩 정품 국내정식 유통제품',
+    'AMD 라이젠7 5세대 7800X3D 라파엘 멀티팩 정품 국내정식 유통제품',
+];
+
+export default data;
