+++ muldum/src/components/Modal/Delete/type.ts
@@ -1,0 +1,4 @@
+export interface Props {
+    onCancel: () => void;
+    onConfirm: () => void;
+}
