+++ muldum/src/components/Modal/Notice/EditSuccess.tsx
@@ -2,12 +2,9 @@ import { useEffect, useRef, useState } from 'react';
 import CheckIcon from '@_assets/modal/check.svg';
 import * as _ from '../style';
 import '@_styles';
+import { Props } from './type'
 
-interface ConfirmDeleteModalProps {
-onClose: () => void;
-}
-
-export default function ConfirmDeleteModal({ onClose }: ConfirmDeleteModalProps) {
+export default function ConfirmDeleteModal({ onClose }: Props) {
 const modalRef = useRef<HTMLDivElement>(null);
 const [counts, setSecondsLeft] = useState(5);
