+++ muldum/src/components/Modal/Notice/type.ts
@@ -1,0 +1,3 @@
+export interface Props {
+onClose: () => void;
+}
