+++ muldum/src/page/Item/Approval/ClubList.ts
@@ -1,0 +1,15 @@
+const Clubs = [
+    '아라',
+    'ODYSSEY',
+    '인서트',
+    '솔빗',
+    'Tera',
+    '하로',
+    'Pluto',
+    'PARADOX',
+    'Echo',
+    '팔레토',
+    '온나미로',
+];
+
+export default Clubs;
