+++ muldum/src/page/Notice/Create/CreateNotice.tsx
@@ -5,16 +5,7 @@ import NavBar from '@_components/NavBar/NavBar';
 import EditSuccess from '@_modal/Notice/EditSuccess'
 import data from '../Detail/data';
 import '@_styles';
-export interface Notice {
-    idx: number;
-    title: string;
-    date: string;
-    formattedDate: string;
-    image: string;
-    imageCaption: string;
-    content: string[];
-    author: string;
-}
+import { Notice } from './type';
 
 export default function CreateNotice() {
     const navigate = useNavigate();
