+++ muldum/src/page/Notice/Detail/Edit/type.ts
@@ -1,0 +1,10 @@
+export interface Notice {
+    idx: number;
+    title: string;
+    date: string;
+    formattedDate: string;
+    image: string;
+    imageCaption: string;
+    content: string[];
+    author: string;
+}
