+++ muldum/src/page/Notice/Notice/type.ts
@@ -1,0 +1,17 @@
+export interface PaginationProps {
+    currentPage: number;
+    totalPages: number;
+    onPageChange: (page: number) => void;
+}
+
+export interface NoticeItem {
+    idx: number;
+    title: string;
+    date: string;
+}
+
+export interface BoxProps {
+    idx: number;
+    title: string;
+    date: string;
+}
